requirements.txt
glob2==0.7
numpy==1.23.5
opencv-python==4.8.0.76
python-dotenv==1.0.0
tensorflow==2.14.0
# tensorflow-gpu>=2.0 
tensorflow-io==0.31.0
wandb==0.15.12
pydot==1.4.2
# windows: install graphviz
tqdm==4.66.1